Factored out Equality[Row]. Set user-provided schema handling to respect nullability via SparkSession config option.
