+++ src/test/scala/cognite/spark/v1/wdl/RowEquality.scala
@@ -1,0 +1,21 @@
+package cognite.spark.v1.wdl
+
+import org.apache.spark.sql.Row
+import org.scalactic.Equality
+
+private[wdl] object RowEquality { // scalastyle:ignore object.name
+  implicit val rowEq: Equality[Row] =
+    new Equality[Row] {
+      def areEqual(a: Row, b: Any): Boolean =
+        b match {
+          case p: Row =>
+            val fieldNames = a.schema.fieldNames
+            if (fieldNames.toSet == p.schema.fieldNames.toSet) {
+              a.getValuesMap(fieldNames) == p.getValuesMap(fieldNames)
+            } else {
+              false
+            }
+          case _ => false
+        }
+    }
+}
